api/requirements.txt
flask-cors
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
requests
gunicorn